+++ app.json
@@ -1,4 +1,4 @@
 {
   "name": "todo_app",
-  "displayName": "todo_app"
+  "displayName": "ToDoNotes"
 }
