+++ package.json
@@ -28,6 +28,7 @@
   },
   "jest": {
     "preset": "react-native",
+    "setupTestFrameworkScriptFile": "<rootDir>/configure-tests.js",
     "transform": {
       "^.+\\.js$": "<rootDir>/node_modules/react-native/jest/preprocessor.js"
     }
