+++ src/App.js
@@ -9,7 +9,7 @@
 import React, {Component} from 'react';
 import {View} from 'react-native';
 import {Provider} from "react-redux";
-import NotesListScreen from "./notes/list/NotesListScreen";
+import NotesListScreen from "notes/notes-list/NotesListScreen";
 import {createStackNavigator} from "react-navigation";
 import NotesDetailScreen from "./notes/detail/NotesDetailScreen";
 import store from "./redux/store";
