+++ src/models.js
@@ -3,4 +3,7 @@ export type Note = {
   title: string,
   description: string
 }
+export type Store = {
+  +notesList: Array<Note>
+}
