+++ src/notes/list/package.json
@@ -1,0 +1,1 @@
+{"name": "notes-list"}
